pyproject.toml
[project]
name = "football-manager"
version = "0.1.0"
description = "Football Team Assignment App"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml",
    "uvicorn>=0.38.0",
    "fasthtml-hf",
]

[tool.ruff.lint]
# Ignore F403 and F405 for star imports from fasthtml.common
# FastHTML is designed to work with star imports (from fasthtml.common import *)
# F403: star import used; unable to detect undefined names
# F405: may be undefined, or defined from star imports
# All HTML components like P, Div, H2, etc. are exported from fasthtml.common
ignore = ["F403", "F405"]
